Move metric helpers into main evaluation module
